fix: Tightened arrow position and improved arrow head
This change just thins the arrow and puts it between the text and icons, it does not introduce a new column.

In general this is an effective fix, but there is one unresolved issue. When we have multiple arrows active at once that overlap it is hard to tell which start point goes with what end point. Not sure of a great solution to this yet.

Could do some sort of color coding, which may just be distracting, or we could allow the arrows to check for each other and try to bend them around one another but I don't think we have the space for that. Still brainstorming solutions, let me know if you have a better idea.