Added Lambda Name Handling for `resolveMethodName` (#450)
* Created a tail-recursive solution for name resolution and added handling for nested lambdas

* Added tests for parentMethodName

* Some refactoring, also fixed call graph logic which assumes 1:1 caller callee relationship

* Ultimately settling on some string formatting of method full name for most language agnostic approach once "normal" method found

* Import update

* Implemented lookahead for anon class constructor

* Patched Anon Class Naming to Match `javap` (#466)

This change adapts how anonymous class names are enumerated to more closely match the Java compiler naming scheme.

One issue now is that the provenance is not clear from the file name alone (we oly get the nearest type). For example, if an anonymous class is within a method, we lose information from which method. The solution would be to traverse the AST.