Persist OOM Flag on `OutOfMemoryException` (#502)
* Persist OOM Flag on `OutOfMemoryException`
`OutOfMemoryException`s can happen in one of many threads in Brokk, thus an `UncaughtExceptionHandler` class called `OomShutdownHandler` is registered to set a "flag" file if such an error occurs. If the JVM permits, an error message is down to the user but this is not guaranteed.

As the JVM is in a fragile state, it was found that creating a file in the Brokk config directory was the most reliable way to persist the message. When the application is started, a warning dialog is displayed and the active sessions are cleared.