Native installers (#605)
* feat: Add native desktop installers to CI

This uses jDeploy to generate native desktop installers on each commit, and each release.

This includes auto-update support so that users don't need to re-install with each release.
Each time the app opens, it checks for the latest from github, and will download if there is a newer version.

* fix jdeploy workflow to get correct jar

* need jar path, not just jar name

* require full jdk to see if that fixes Error loading embedding model

* fix add-modules directive

* Added install splash image

* moved jdeploy release builds to release workflow

changed jdeploy workflow to only generate branch-based installers, and only on master branch

* fixed license in package.json