Compositional Analyzer Capability Checks (#708)
A number of search tools associate certain capabilities of tools to the CPG, when other analyzers such as the TreeSitter analyzers can provide these themselves. I suggest we add interfaces which dictate the features an analyzer may be capable of.

This would make the `IAnalyzer` API cleaner, throw fewer `UnimplementedException`s, and keep the polymorphism flat. #selling

The changes are as follows:
* Migrated certain API calls from `IAnalyzer` to `CallGraphProvider`, `SkeletonProvider`, `SourceCodeProvider`, `UsagesProvider`.
* Removed `isCpg` from `IAnalyzer`. Callers of this function were refactored to check if the analyser implements one of the following interfaces depending on the API it requires.
* Various analyser users can now make use of analysers that partially support historically "CPG only" capabilities if they implement the above. 

Resolves https://github.com/BrokkAi/brokk/issues/705