Separated Completion Task from Wrangling `searchDefinitions` to Dedicated `autocompleteDefinitions` (#839)
This approach tries something new, that is, using a navigable set structure to find initial auto-complete candidates. It involves introducing a new global data structure which is not ideal, but further makes the case for something like Chronicle map/ a database down the line.

This is backed by a skip list and provides us with `NavigableSet` methods like `tailSet` which are used if possible to speed up prefix-based queries. Other queries that are more camel-case or hierarchical leverage a more expensive, O(N), strategy.

Symbols are also cached to matching code units for fast lookups.