Normalize `$` Class Separators as `.` (#1044)
* Normalize `$` Class Separators as `.`
A problem with using the `$` separators is that JDT does not use them. It is easy to go from TreeSitter names that have dollar signs, but not the other way around.

This may mean that any `fqNames` from JDT referring to nested classes cannot be resolved.

This PR ensures TreeSitter analyzers use `.` separators for inner classes to be consistent with JDT's naming scheme. This simplifies a few checks elsewhere too.

* Fix tests