fix: Make max memory configurable at runtime in jDeploy
This creates a config file that will be picked up by jDeploy at runtime to override JVM arguments.
It also updates the jDeploySettingsUtil to update this file instead of the old ~/.brokk/jdeploy.args file.
It also updates the release workflow to use jdeploy 5.2.0-dev.4 which includes support for this runtime config file.
Will use this for testing, and will create new jdeploy release and remove this dev version dependency if all goes well.