fix: Make max memory configurable at runtime in jDeploy (#1264)
This creates a config file that will be picked up by jDeploy at runtime to override JVM arguments.
It also updates the jDeploySettingsUtil to update this file instead of the old ~/.brokk/jdeploy.args file.
It also updates the release workflow to use jdeploy 5.2.0-dev.4 which includes support for this runtime config file.

fixed package.json