Separate Constructor/Type Menu Items on PreviewTextPanel (#1405)
I observed a bug where only the first of either the constructor or class' CodeUnit will be behind the menu items of PreviewTextPanel.

Now, if the item matched is a class, it will include the constructor as an additional item with a check for source code before adding the source-code equivalent.