Fix/jdeploy snapshots (#1570)
* fix: Snapshot releases

Previously it created jdeploy snapshots on the master branch directly, which creates a master tag, but this was conflicting with git pulls for origin master, as it would sometimes pull the tag instead of the branch.
With this change, it will sync the master branch with another branch called master-snapshot, and the jdeploy bundles will be run on that master-snapshot branch instead.

* explicitly trigger jdeploy workflow on master-snapshot branch

* added permission to workflow trigger

* updated development.md with new development release URL