Allow Max Concurrency to be Specified for AdaptiveExecutor (#1589)
Given that the task of disambiguating usages can quickly exceed a sensible amount of requests to LLM providers, this PR allows the FuzzyUsageFinder/RelevanceClassifiers to provide a maximum amount of concurrency. This should hopefully prevent the "too many concurrent streams" warning from popping up as often as it does for some users now.

Misc: Fix ordering issue that documentation promises