fix: LocalFileRepo hangs on large repos
This adds a tracked file cache to the LocalFileRepo, just like the IGitRepo.
If you open a project with lots of files that doesn't use git, then it would just hang because it would be crawling the entire directory for each node of the file tree, on the EDT.  This fixes that.